fix(custom-error): fix imports after custom error file rename
