+++ src/shared/helpers/index.ts
@@ -1,7 +1,7 @@
 export * from './clipboard';
 export * from './dropdown';
 export * from './env';
-export * from './error';
+export * from './custom-error';
 export * from './formatters';
 export * from './gql';
 export * from './link';
